+++ examples/pipeline.php
@@ -11,6 +11,7 @@
 
 /**
  * Pipeline
+ * Note. Be careful in use pipeline with Clusters.
  */
 
 namespace Examples;