+++ tests/Integration/Version3x0/ClusterTest.php
@@ -12,6 +12,9 @@ namespace Test\Integration\Version3x0;
 
 use RedisClient\Client\AbstractRedisClient;
 use RedisClient\Cluster\ClusterMap;
+use RedisClient\Exception\CrossSlotResponseException;
+use RedisClient\Exception\MovedResponseException;
+use RedisClient\Pipeline\Pipeline;
 
 include_once(__DIR__ . '/../ClusterVersionTest.php');
 
@@ -158,6 +161,56 @@ class ClusterTest extends \Test\Integration\ClusterVersionTest {
         foreach ($this->getServers() as $server) {
             $this->assertSame(true, 0 < strpos($nodes, $server));
         }
+    }
+
+    /**
+     * @see \RedisClient\Command\Traits\Version3x0\ClusterCommandsTrait::clusterNodes
+     */
+    public function test_validPipeline() {
+        $Redis = $this->getRedisClient();
+        $result = $Redis->pipeline(function($Pipeline) {
+            /** @var Pipeline $Pipeline */
+            $Pipeline->set('user{42}', 'Alexander');
+            $Pipeline->set('country{42}', 'UK');
+            $Pipeline->set('city{42}', 'London');
+            $Pipeline->mget(['user{42}', 'country{42}', 'city{42}']);
+        });
+
+        $this->assertSame(true, is_array($result));
+        $this->assertSame(4, count($result));
+        $this->assertSame(true, $result[0]);
+        $this->assertSame(true, $result[1]);
+        $this->assertSame(true, $result[2]);
+        $this->assertSame('Alexander', $result[3][0]);
+        $this->assertSame('UK', $result[3][1]);
+        $this->assertSame('London', $result[3][2]);
+    }
+
+    /**
+     * @see \RedisClient\Command\Traits\Version3x0\ClusterCommandsTrait::clusterNodes
+     */
+    public function test_invalidPipeline() {
+        $Redis = $this->getRedisClient();
+        $result = $Redis->pipeline(function($Pipeline) {
+            /** @var Pipeline $Pipeline */
+            $Pipeline->set('user{42}', 'Alexander');
+            $Pipeline->set('user{43}', 'Irina');
+            $Pipeline->set('user{44}', 'Aram');
+            $Pipeline->mget(['user{42}', 'user{43}', 'user{44}']);
+        });
+
+        $this->assertSame(true, is_array($result));
+        $this->assertSame(4, count($result));
+        $this->assertSame(true, $result[0]);
+        $this->assertSame(true, $result[1] instanceof MovedResponseException);
+        $this->assertSame(true, $result[2] instanceof MovedResponseException);
+        $this->assertSame(true, $result[3] instanceof CrossSlotResponseException);
+
+        $this->assertSame('127.0.0.1:7001', $result[1]->getServer());
+        $this->assertSame(3937, $result[1]->getSlot());
+        $this->assertSame('127.0.0.1:7003', $result[2]->getServer());
+        $this->assertSame(16262, $result[2]->getSlot());
+        $this->assertSame('CROSSSLOT Keys in request don\'t hash to the same slot', $result[3]->getMessage());
     }
 
 }